V1/BE/src/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
jinja2
